refactor: split deploy tasks
