+++ deploy/modes.py
@@ -1,0 +1,5 @@
+"""
+Allowed deployment modes
+"""
+DEVELOPMENT_MODE = 'development'
+PRODUCTION_MODE = 'production'
