+++ deploy/tasks/service_lifting.py
@@ -1,0 +1,12 @@
+"""
+TODO
+"""
+from ansible import AnsibleExecutor
+
+
+async def configure_nginx(ansible: AnsibleExecutor) -> None:
+    """ Configures nginx
+    Args:
+        ansible: instance of ansible executor
+    """
+    pass
